+++ reflex_django_tests/test_page_on_load.py
@@ -9,7 +9,7 @@ from reflex_django.conf import configure_django
 configure_django()
 
 from reflex_django.app_factory import reset_app_factory_cache, sync_page_load_events  # noqa: E402
-from reflex_django.decorators import clear_page_registry, page
+from reflex_django.pages.decorators import clear_page_registry, page
 from reflex_django.states import AppState